// src/app/core/models/useCases/identity-tenant/common/filtering-operation.ts
export enum FilteringOperation {
  None = 'None',
  Equals = 'Equals',
  Contains = 'Contains',
  StartsWith = 'StartsWith',
  EndsWith = 'EndsWith',
  NotEquals = 'NotEquals',
  GreaterThan = 'GreaterThan',
  GreaterThanOrEqual = 'GreaterThanOrEqual',
  LessThan = 'LessThan',
  LessThanOrEqual = 'LessThanOrEqual',
  In = 'In',
  NotIn = 'NotIn',
  Between = 'Between',
  NotBetween = 'NotBetween',
  IsNull = 'IsNull',
  IsNotNull = 'IsNotNull'
}